Interview

+++ Algorithm/src/cisco/CountDigitOne.java
@@ -9,7 +9,7 @@ public class CountDigitOne {
         int count = 0;
         for (long i = 1; i <= n; i *= 10) {
             long divider = i * 10;
-            count += (n / divider) * i + min(max(n % divider - i + 1, 0), i);
+            count += (int) ((n / divider) * i + min(max(n % divider - i + 1, 0), i));
         }
         return count;
     }
